+++ src/lib.rs
@@ -40,7 +40,8 @@ pub mod validate;
 mod wasm;
 
 pub use crate::core::{
-    detect, from_xml, to_xml, to_xml_fragment, BusinessArea, Error, MxId, MxMessage, Result,
+    detect, from_xml, read_business_message, to_xml, to_xml_fragment, BusinessArea,
+    BusinessMessage, Error, MxId, MxMessage, Result,
 };
 #[cfg(feature = "serde")]
 pub use crate::core::{from_json, to_json};